Beginning reorganization of OutputFile class
Making ReadOutputFile and WriteOutputFile to avoid bloating
a single class
