+++ src/filetypes.h
@@ -44,8 +44,8 @@ enum OutputFormat {
   OutputNetcdf3 = 0x003,
   /// netCDF4 formatted Adcirc output
   OutputNetcdf4 = 0x004,
-  /// Xdmf formatted Adcirc output [not implemented]
-  OutputXdmf = 0x005,
+  /// Hdf5 formatted Adcirc output [not implemented]
+  OutputHdf5 = 0x005,
   /// Unknown output format
   OutputUnknown = 0x999
 };
@@ -54,7 +54,6 @@ bool checkFiletypeAsciiFull(const std::string &filename);
 bool checkFiletypeAsciiSparse(const std::string &filename);
 bool checkFiletypeNetcdf3(const std::string &filename);
 bool checkFiletypeNetcdf4(const std::string &filename);
-bool checkFiletypeXdmf(const std::string &filename);
 bool inquireNetcdfFormat(const std::string &filename, int &format);
 
 }  // namespace Output
