+++ src/writeoutputfile.cpp
@@ -1,0 +1,6 @@
+#include "writeoutputfile.h"
+
+WriteOutputFile::WriteOutputFile()
+{
+
+}
