+++ src/writeoutputfile.h
@@ -1,0 +1,11 @@
+#ifndef WRITEOUTPUTFILE_H
+#define WRITEOUTPUTFILE_H
+
+
+class WriteOutputFile
+{
+public:
+  WriteOutputFile();
+};
+
+#endif // WRITEOUTPUTFILE_H
