+++ swig/adcirc.i
@@ -47,6 +47,7 @@
 #include "attributemetadata.h"
 #include "nodalattributes.h"
 #include "outputfile.h"
+#include "readoutputfile.h"
 #include "outputrecord.h"
 #include "harmonicsrecord.h"
 #include "harmonicsoutput.h"
@@ -102,6 +103,7 @@ namespace std {
 %include "attributemetadata.h"
 %include "nodalattributes.h"
 %include "outputfile.h"
+%include "readoutputfile.h"
 %include "outputrecord.h"
 %include "harmonicsrecord.h"
 %include "harmonicsoutput.h"
