+++ testing/cxx_tests/cxx_readasciivector.cpp
@@ -24,8 +24,8 @@
 int main() {
   using namespace Adcirc::Output;
 
-  std::unique_ptr<OutputFile> output(
-      new OutputFile("test_files/sparse_fort.64"));
+  std::unique_ptr<ReadOutputFile> output(
+      new ReadOutputFile("test_files/sparse_fort.64"));
 
   //...Open file
   output->open();
