+++ testing/cxx_tests/cxx_readmaxele.cpp
@@ -23,7 +23,7 @@
 int main() {
   using namespace Adcirc::Geometry;
   using namespace Adcirc::Output;
-  std::unique_ptr<OutputFile> output(new OutputFile("test_files/maxele.63"));
+  std::unique_ptr<ReadOutputFile> output(new ReadOutputFile("test_files/maxele.63"));
 
   //...Open file
   output->open();
