+++ testing/cxx_tests/cxx_readnetcdfvector.cpp
@@ -23,7 +23,7 @@
 int main() {
   using namespace Adcirc::Geometry;
   using namespace Adcirc::Output;
-  std::unique_ptr<OutputFile> output(new OutputFile("test_files/fort.64.nc"));
+  std::unique_ptr<ReadOutputFile> output(new ReadOutputFile("test_files/fort.64.nc"));
   output->open();
   output->read();
   output->read();
